Implement Shalizi merging pseudocode.
I implemented the pseudocode algorithm for performing state merging
(via hypothesis testing) as described in all of Shalizi's papers
related to spatial computational mechanics.

This is working wrt a Bernoulli process. With the even process, I
haven't figured out how to get it to distinguish between, say, the
distribution of '111' when proceeded by a '1' vs. by a '0.'