Fixed a bug in stf_methods.py.
Originally, sft_methods had a place where I manipulated one of the
lists passed as an argument to csmr(). Since it was a list of lists,
and the internal lists were passed by reference, they got manipulated
by csmr, which was not what I expected. I fixed this by using
copy.deepcopy() from the copy module.